split jirarepository by context
