--- cmd/api/BUILD.bazel
@@ -1,29 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
-load("@io_bazel_rules_docker//container:container.bzl", "container_image")
-
-go_library(
-    name = "api_lib",
-    srcs = ["main.go"],
-    importpath = "github.com/stebennett/squad-dashboard/cmd/api",
-    visibility = ["//visibility:private"],
-    deps = [
-        "//pkg/api/controllers",
-        "//pkg/api/routes",
-        "//pkg/jirarepository",
-        "//pkg/jirastatsservice",
-        "//vendor/github.com/gorilla/mux",
-    ],
-)
-
-go_binary(
-    name = "api",
-    embed = [":api_lib"],
-    visibility = ["//visibility:public"],
-)
-
-container_image(
-    name = "image",
-    base = "@alpine_linux_amd64//image",
-    entrypoint = ["/api"],
-    files = [":api"],
-)
