+++ cmd/jiraissuecalculator/BUILD.bazel
@@ -8,6 +8,8 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//cmd/jiraissuecalculator/calculator",
+        "//pkg/jiracalculationsrepository",
+        "//pkg/jiraconfigrepository",
         "//pkg/jirarepository",
         "//vendor/github.com/Netflix/go-env",
     ],
