--- pkg/api/controllers/BUILD.bazel
@@ -1,17 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "controllers",
-    srcs = [
-        "health_controller.go",
-        "jira_data_controller.go",
-        "jira_stats_controller.go",
-    ],
-    importpath = "github.com/stebennett/squad-dashboard/pkg/api/controllers",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/jirarepository",
-        "//pkg/jirastatsservice",
-        "//vendor/github.com/gorilla/mux",
-    ],
-)
