--- pkg/api/controllers/health_controller.go
@@ -1,11 +1,0 @@
-package controllers
-
-import (
-	"encoding/json"
-	"net/http"
-)
-
-func HealthHandler(w http.ResponseWriter, r *http.Request) {
-	w.WriteHeader(http.StatusOK)
-	json.NewEncoder(w).Encode(map[string]bool{"ok": true})
-}
