--- pkg/api/routes/health_routes.go
@@ -1,10 +1,0 @@
-package routes
-
-import (
-	"github.com/gorilla/mux"
-	"github.com/stebennett/squad-dashboard/pkg/api/controllers"
-)
-
-func HealthRoutes(router *mux.Router) {
-	router.HandleFunc("/health", controllers.HealthHandler).Methods("GET")
-}
