--- pkg/api/routes/jira_data_routes.go
@@ -1,10 +1,0 @@
-package routes
-
-import (
-	"github.com/gorilla/mux"
-	"github.com/stebennett/squad-dashboard/pkg/api/controllers"
-)
-
-func JiraDataRoutes(dataController controllers.JiraDataController, router *mux.Router) {
-	router.HandleFunc("/jira/projects", dataController.ListProjects).Methods("GET")
-}
