+++ pkg/jiraconfigrepository/BUILD.bazel
@@ -1,0 +1,11 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "jiraconfigrepository",
+    srcs = [
+        "jiraconfigrepository.go",
+        "postgresjiraconfigrespository.go",
+    ],
+    importpath = "github.com/stebennett/squad-dashboard/pkg/jiraconfigrepository",
+    visibility = ["//visibility:public"],
+)
