--- pkg/jirastatsservice/BUILD.bazel
@@ -1,14 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "jirastatsservice",
-    srcs = ["jirastatsservice.go"],
-    importpath = "github.com/stebennett/squad-dashboard/pkg/jirastatsservice",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/dateutil",
-        "//pkg/jirarepository",
-        "//pkg/statsmodels",
-        "//pkg/statsutil",
-    ],
-)
